# docs/api_ml_service/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Cripto Service API",
        "description": "API for training and predicting cryptocurrency prices using machine learning models.",
        "version": "1.0.0"
    },
    "paths": {
        "/models/train": {
            "post": {
                "tags": [
                    "Models"
                ],
                "summary": "Train Model",
                "operationId": "train_model_models_train_post",
                "parameters": [
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TrainRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TrainResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/models/predict": {
            "post": {
                "tags": [
                    "Models"
                ],
                "summary": "Predict",
                "operationId": "predict_models_predict_post",
                "parameters": [
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PredictRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PredictionResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/models/predict/hour": {
            "post": {
                "tags": [
                    "Models"
                ],
                "summary": "Predict For Hour",
                "operationId": "predict_for_hour_models_predict_hour_post",
                "parameters": [
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PredictHourRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PredictionHourResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/models/": {
            "get": {
                "tags": [
                    "Models"
                ],
                "summary": "Get All Models",
                "operationId": "get_all_models_models__get",
                "parameters": [
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ModelResponse"
                                    },
                                    "title": "Response Get All Models Models  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/models/symbol/{symbol}": {
            "get": {
                "tags": [
                    "Models"
                ],
                "summary": "Get Models By Symbol",
                "operationId": "get_models_by_symbol_models_symbol__symbol__get",
                "parameters": [
                    {
                        "name": "symbol",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Symbol"
                        }
                    },
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ModelResponse"
                                    },
                                    "title": "Response Get Models By Symbol Models Symbol  Symbol  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/models/active/{symbol}": {
            "get": {
                "tags": [
                    "Models"
                ],
                "summary": "Get Active Model By Symbol",
                "operationId": "get_active_model_by_symbol_models_active__symbol__get",
                "parameters": [
                    {
                        "name": "symbol",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Symbol"
                        }
                    },
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ModelResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/models/active": {
            "get": {
                "tags": [
                    "Models"
                ],
                "summary": "Get All Active Models By Symbol",
                "operationId": "get_all_active_models_by_symbol_models_active_get",
                "parameters": [
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ModelResponse"
                                    },
                                    "title": "Response Get All Active Models By Symbol Models Active Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/models/id/{model_id}": {
            "get": {
                "tags": [
                    "Models"
                ],
                "summary": "Get Model By Id",
                "operationId": "get_model_by_id_models_id__model_id__get",
                "parameters": [
                    {
                        "name": "model_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Model Id"
                        }
                    },
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ModelResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "ModelResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "model_name": {
                        "type": "string",
                        "title": "Model Name"
                    },
                    "symbol": {
                        "type": "string",
                        "title": "Symbol"
                    },
                    "model_algorithm": {
                        "type": "string",
                        "title": "Model Algorithm"
                    },
                    "model_version": {
                        "type": "string",
                        "title": "Model Version"
                    },
                    "mae": {
                        "type": "number",
                        "title": "Mae"
                    },
                    "rmse": {
                        "type": "number",
                        "title": "Rmse"
                    },
                    "observations": {
                        "type": "integer",
                        "title": "Observations"
                    },
                    "active": {
                        "type": "boolean",
                        "title": "Active"
                    },
                    "model_path": {
                        "type": "string",
                        "title": "Model Path"
                    },
                    "training_date": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Training Date"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "model_name",
                    "symbol",
                    "model_algorithm",
                    "model_version",
                    "mae",
                    "rmse",
                    "observations",
                    "active",
                    "model_path",
                    "training_date"
                ],
                "title": "ModelResponse"
            },
            "PredictHourRequest": {
                "properties": {
                    "symbol": {
                        "type": "string",
                        "title": "Symbol"
                    },
                    "hour": {
                        "type": "integer",
                        "title": "Hour"
                    }
                },
                "type": "object",
                "required": [
                    "symbol",
                    "hour"
                ],
                "title": "PredictHourRequest"
            },
            "PredictRequest": {
                "properties": {
                    "symbol": {
                        "type": "string",
                        "title": "Symbol"
                    }
                },
                "type": "object",
                "required": [
                    "symbol"
                ],
                "title": "PredictRequest"
            },
            "PredictionHourResponse": {
                "properties": {
                    "symbol": {
                        "type": "string",
                        "title": "Symbol"
                    },
                    "current_price": {
                        "type": "number",
                        "title": "Current Price"
                    },
                    "model": {
                        "type": "string",
                        "title": "Model"
                    },
                    "version": {
                        "type": "string",
                        "title": "Version"
                    },
                    "prediction": {
                        "$ref": "#/components/schemas/PredictionItemResponse"
                    }
                },
                "type": "object",
                "required": [
                    "symbol",
                    "current_price",
                    "model",
                    "version",
                    "prediction"
                ],
                "title": "PredictionHourResponse"
            },
            "PredictionItemResponse": {
                "properties": {
                    "prediction_hour": {
                        "type": "integer",
                        "title": "Prediction Hour"
                    },
                    "estimated_price": {
                        "type": "number",
                        "title": "Estimated Price"
                    },
                    "percentage_variation": {
                        "type": "number",
                        "title": "Percentage Variation"
                    }
                },
                "type": "object",
                "required": [
                    "prediction_hour",
                    "estimated_price",
                    "percentage_variation"
                ],
                "title": "PredictionItemResponse"
            },
            "PredictionResponse": {
                "properties": {
                    "symbol": {
                        "type": "string",
                        "title": "Symbol"
                    },
                    "current_price": {
                        "type": "number",
                        "title": "Current Price"
                    },
                    "model": {
                        "type": "string",
                        "title": "Model"
                    },
                    "version": {
                        "type": "string",
                        "title": "Version"
                    },
                    "predictions": {
                        "items": {
                            "$ref": "#/components/schemas/PredictionItemResponse"
                        },
                        "type": "array",
                        "title": "Predictions"
                    }
                },
                "type": "object",
                "required": [
                    "symbol",
                    "current_price",
                    "model",
                    "version",
                    "predictions"
                ],
                "title": "PredictionResponse"
            },
            "TrainRequest": {
                "properties": {
                    "symbol": {
                        "type": "string",
                        "title": "Symbol"
                    },
                    "coin_gecko_id": {
                        "type": "string",
                        "title": "Coin Gecko Id"
                    }
                },
                "type": "object",
                "required": [
                    "symbol",
                    "coin_gecko_id"
                ],
                "title": "TrainRequest"
            },
            "TrainResponse": {
                "properties": {
                    "trained": {
                        "type": "boolean",
                        "title": "Trained"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "symbol": {
                        "type": "string",
                        "title": "Symbol"
                    },
                    "last_training": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Last Training"
                    },
                    "new_records": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "New Records"
                    },
                    "hours_since_training": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Hours Since Training"
                    },
                    "observations": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Observations"
                    },
                    "mae": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Mae"
                    },
                    "rmse": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Rmse"
                    },
                    "model_path": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Model Path"
                    }
                },
                "type": "object",
                "required": [
                    "trained",
                    "message",
                    "symbol"
                ],
                "title": "TrainResponse"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}